requirements.txt

viam-sdk==0.42.0

typing-extensions
pillow
